backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
colorama==0.4.6
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
docopt==0.6.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
model-bakery==1.17.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.4.0
psycopg==3.1.17
psycopg-binary==3.1.17
psycopg-pool==3.2.1
pydantic==2.6.0
pydantic_core==2.16.1
PySocks==1.7.1
pytest==8.0.0
pytest-django==4.8.0
pytest-watch==4.2.0
pytz==2023.4
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
watchdog==3.0.0
wsproto==1.2.0
